Devide codes into helper and utils, and add some test codes.
