+++ Unicoen.Apps/RefactoringDSL.Tests/FindUtilTest.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Unicoen.Apps.RefactoringDSL.Tests {
+	class FindUtilTest {
+	}
+}
