+++ Unicoen.Apps/RefactoringDSL/Unicoen.Apps.RefactoringDSL.csproj
@@ -48,6 +48,7 @@
     <Compile Include="EncapsulateField.cs" />
     <Compile Include="EncapsulateFieldHelper.cs" />
     <Compile Include="IRefactorable.cs" />
+    <Compile Include="Java\EncapsulateCollectionHelperForJava.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Util\FindUtil.cs" />
